feat: Using Navigation Listeners
You can prevent the drawer from being swiped anywhere other than the startDestination. All we need to do is call addOnDestinationChangedListener with a lambda that sets the DrawerLockMode depending on what destination we’re navigating to. When the id of our NavDestination matches the startDestination of our graph, we’ll unlock the drawerLayout; otherwise, we’ll lock and close the drawerLayout.